Avoid eval() entirely, part 1
