+++ xcube/core/varexpr/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
+from .context import VarExprContext
+from .error import VarExprError
+from .helpers import split_var_assignment
+from .varexpr import VarExpr
+from .varexpr import evaluate
+from .varexpr import parse
