+++ xcube/core/varexpr/error.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
+
+class VarExprError(ValueError):
+    """Exception raised by code in the ``xcube.code.varexpr` module."""
